pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.106.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pymupdf>=1.26.4",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "uvicorn[standard]>=0.35.0",
]
